refactor: 1.迁移 CustomControl 目录结构  2.补充 Control 文档
