+++ src/components/Control/CustomControl/types.ts
@@ -1,10 +1,10 @@
 import type { IControlOption } from '@antv/l7';
-import type { CommonProps } from '../../types';
+import type { CommonProps } from '../../../types';
 
 /**
  * 组件类型定义
  */
 export interface CustomControlProps extends CommonProps, Omit<Partial<IControlOption>, 'style'> {
-  /** 子组件 */
+  /** 控件容器包含的内容 */
   children?: React.ReactNode;
 }
