+++ src/types/common.ts
@@ -1,8 +1,8 @@
 import type { CSSProperties } from 'react';
 
 export type CommonProps = {
-  /** 类名 */
+  /** 容器类名 */
   className?: string;
-  /** 行内样式 */
+  /** 容器内敛样式 */
   style?: CSSProperties;
 };
